requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
djangorestframework==3.14.0
geopy==2.4.0
routingpy==1.3.0
psycopgr==1.0.6
django-cors-headers==4.3.1
